refactor(commands/nostr): move commands to their own file
